Sketching out a "Classes" and "Schools" Page, + new CTA Include (#44)
This PR sets the groundwork for two new pages:

* a "classes" page, which lists all the classes we've ever taught, including learning goals, lead instructors, and links for more information
* a "schools" page, where you can see all of our partnered schools

For now, I've finalized all of the content I want on these two pages (though the content they link to, like course outlines & learning paths, are still WIP). We'll have to revisit schools later when the school year starts.

In addition to those two new pages, I've also created a Jekyll include, `contact-cta.html`. This is copy (the links which we'll update soon) intended to pad out the space between the last content on a page and the footer, with a CTA to contact us. LMK what you think.

Finally, I made several small site-wide design changes:

* icons inside of buttons now have an explicit margin-right to guarantee spacing
* `line-height` increased to 1.25 for <p> and 1.5 for ul > li for better readability