A More Informative 404 Page (#51)
This is a short and sweet PR that upgrades our 404 page: we've added an illustration, created a set of CTAs for users to report problems, and included a Spanish translation (as we might start adding Spanish content soon, @ Brockton). Big thanks to Evan Zhong (@evazhog) , Nhi Van (@vohndernet), and Ashley Ghodsian for helping us out!