Adds images for each blog post (#156)
This branch adds a default image associated with each blog post. This image is shown on the post, the main blog page, and all blog page recommendation. This branch also restructures the way images are stored for posts, with each post having its own folder to contain its images. The default image is named default, and the path to it is stored in the post .md file.

Closes #128

Co-authored-by: Matt Wang <matt@matthewwang.me>